Permission to data according Roles

+++ Presentation/Real_Estate.Presentation.WebApp/Controllers/AdminController.cs
@@ -38,6 +38,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
         }
         public async Task<IActionResult> Index()
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             HomeAdminViewModel homeAdminViewModel = new();
             homeAdminViewModel = await _userService.GetUsersQuantity();
             var propertiesVM = await _propertiesService.GetAllWithData();
@@ -48,6 +65,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public async Task<IActionResult> AgentsList()
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             var usersList = await _userService.GetAllUsersViewModels();
 
             List<UserViewModel> AgentsList = usersList.Where(user => 
@@ -66,6 +100,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public async Task<IActionResult> AdminManager()
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             var usersList = await _userService.GetAllUsersViewModels();
             List<UserViewModel> adminList = usersList.Where(
                 user => user.Role == Roles.Admin.ToString()).ToList();
@@ -75,6 +126,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public async Task<IActionResult> DevsManager()
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             var usersList = await _userService.GetAllUsersViewModels();
             List<UserViewModel> adminList = usersList.Where(
                 user => user.Role == Roles.Developer.ToString()).ToList();
@@ -84,6 +152,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public IActionResult RegisterUser(string role)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             if (role == "admin")
             {
                 ViewBag.Role = Roles.Admin.ToString();
@@ -102,6 +187,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> RegisterUser(UpdateUserViewModel vm, string role)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             if (!ModelState.IsValid)
             {
                 ViewBag.Role = role;
@@ -153,6 +255,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public async Task<IActionResult> UpdateUser(string username, string role)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             UpdateUserViewModel userSaveViewModel = await _userService.GetUserSaveViewModelByUsername(username);
 
             ViewBag.role = null;
@@ -164,6 +283,24 @@ namespace Real_Estate.Presentation.WebApp.Controllers
 
         public async Task<IActionResult> UpdateUser(UpdateUserViewModel userSaveViewModel, string role)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
+
             if (!ModelState.IsValid)
             {
                 ViewBag.Role = null;
@@ -211,6 +348,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> ChangeUserStatus(string id, string Role)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             var response = await _accountService.ChangeUserStatusAsync(id);
 
             if (Role == "Developer")
@@ -230,6 +384,23 @@ namespace Real_Estate.Presentation.WebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(string id)
         {
+            if (currentlyUser.Roles.FirstOrDefault() == Roles.Agent.ToString())
+            {
+                return RedirectToRoute(new { controller = "Agent", action = "Index" });
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Client.ToString())
+            {
+                return RedirectToRoute(new { controller = "Client", action = "Index" });
+
+            }
+
+            else if (currentlyUser.Roles.FirstOrDefault() == Roles.Developer.ToString())
+            {
+                return RedirectToRoute(new { controller = "Home", action = "AccessDenied" });
+
+            }
+
             var response = await _accountService.DeleteUserAsync(id);
 
             return RedirectToRoute(new { controller = "Admin", action = "AgentsList" });
